[BUGFIX] Add very stupid circumvention of query parameter problem
The Nodeindex has a problem in combination with the RequestService of
Flowpack.ElasticSerach, which means that an trailing ? will be appended
to all qeeries. This does now work very well when adding ?scroll=1m to
the querystring, since ES parsed 1m? as unparsable. To "fix" this, we just
add a dummy parameter instead of fixing the real issue in FlowPack.ElasticSerach